refactor bluehawk w/o uncomment:
